Remove AlbumArtRef field from Track
It couldn't actually be used to link to album art.  The error is probably
in that the Origin and Store fields of internal/locker_proto/data.ImageRef
are missing from ImageRef, but I'd rather not introduce yet more enums to
the package on the off chance that it could work.  Better to wait until
Simon or someone develops a way to upload embedded album art to Google Play.