Add CheckImportResponse function
Uploading a track can fail even with a non-2xx status code, so it's good
to expose a function that parses the response to an upload and checks it
for an error value.  As an added convenience, the function also parses
and returns the uploaded track's server ID for the user.  (The server ID
is actually also embedded in the upload URL, but for symmetry, I think
it makes sense to expose it only on a successful upload.)